refactor(session): extract the debug launcher into src/session/launch/debug-launcher.ts
startDebugging, restartDebugging (with its restartingSessions reentrancy
guard) and waitForDryRunCompletion move verbatim onto DebugLauncher, wired as
DebugLauncher(ctx, proxyLauncher, breakpoints) over a new LaunchContext slice
of OperationsContext. The inline readiness promise inside startDebugging
becomes waitForLaunchReadiness() in launch/launch-readiness.ts — same body,
same eight log lines, still never rejects and still reads session.proxyManager
lazily. restartDebugging calls the launcher's own startDebugging. The facade
keeps one-line delegates for startDebugging and restartDebugging.

waitForDryRunCompletion is public on the launcher; the seven coverage-test
sites move from `(ops as any).waitForDryRunCompletion` to
`internals(ops).launcher.waitForDryRunCompletion`.
