+++ src/session/session-manager-operations.ts
@@ -9,13 +9,11 @@ import {
   FunctionBreakpoint,
   SessionState,
   SessionLifecycleState,
-  sanitizePayloadForLogging,
   type ExceptionBreakMode
 } from '@debugmcp/shared';
-import { ManagedSession, ToolchainValidationState } from './session-store.js';
+import { ManagedSession } from './session-store.js';
 import { DebugProtocol } from '@vscode/debugprotocol';
 import { ErrorMessages } from '../utils/error-messages.js';
-import { checkLaunchToolchain } from '../utils/language-availability.js';
 import { SessionManagerData } from './session-manager-data.js';
 import type { OperationsContext } from './operations-context.js';
 import { BreakpointController } from './breakpoints/breakpoint-controller.js';
@@ -33,15 +31,10 @@ import {
   type ExposeSessionResult,
   type UnexposeSessionResult
 } from './mirror/mirror-controller.js';
-import { reresolveAnchors } from './breakpoints/anchor-resolution.js';
-import {
-  buildLogpointDowngradeLaunchWarning,
-  buildUnboundBreakpointExitWarning
-} from './breakpoints/launch-warnings.js';
 import { ProxyLauncher } from './launch/proxy-launcher.js';
+import { DebugLauncher } from './launch/debug-launcher.js';
 import { logProxyFailure } from './launch/proxy-failure-diagnostics.js';
 import { CustomLaunchRequestArguments, DebugResult } from './session-manager-core.js';
-import { McpError } from '@modelcontextprotocol/sdk/types.js';
 
 /** Result type for evaluate expression operations. */
 export type { EvaluateResult } from './inspection/expression-evaluator.js';
@@ -156,68 +149,12 @@ export abstract class SessionManagerOperations extends SessionManagerData {
   protected readonly hotSwap = new RedefineClassesController(this.opsContext, this.breakpoints);
   protected readonly mirror = new MirrorController(this.opsContext);
   protected readonly proxyLauncher = new ProxyLauncher(this.opsContext);
+  protected readonly launcher = new DebugLauncher(this.opsContext, this.proxyLauncher, this.breakpoints);
 
   /**
-   * Helper method to wait for dry run completion with timeout
+   * Start (or dry-run) a launch-mode debug session. Delegates to the launcher,
+   * which owns the launch sequence end to end.
    */
-  private async waitForDryRunCompletion(
-    session: ManagedSession,
-    timeoutMs: number
-  ): Promise<boolean> {
-    if (session.proxyManager?.hasDryRunCompleted?.()) {
-      this.logger.info(
-        `[SessionManager] Dry run already marked complete for session ${session.id} before wait`
-      );
-      return true;
-    }
-
-    let handler: (() => void) | null = null;
-    let timeoutId: NodeJS.Timeout | null = null;
-
-    try {
-      return await Promise.race([
-        new Promise<boolean>((resolve) => {
-          handler = () => {
-            this.logger.info(
-              `[SessionManager] Dry run completion event received for session ${session.id}`
-            );
-            resolve(true);
-          };
-          this.logger.info(
-            `[SessionManager] Setting up dry-run-complete listener for session ${session.id}`
-          );
-          session.proxyManager?.once('dry-run-complete', handler);
-        }),
-        new Promise<boolean>((resolve) => {
-          timeoutId = setTimeout(() => {
-            if (session.proxyManager?.hasDryRunCompleted?.()) {
-              this.logger.info(
-                `[SessionManager] Dry run marked complete during timeout window for session ${session.id}`
-              );
-              resolve(true);
-              return;
-            }
-            this.logger.warn(
-              `[SessionManager] Dry run timeout after ${timeoutMs}ms for session ${session.id}`
-            );
-            resolve(false);
-          }, timeoutMs);
-        }),
-      ]);
-    } finally {
-      // Clean up immediately
-      if (handler && session.proxyManager) {
-        this.logger.info(
-          `[SessionManager] Removing dry-run-complete listener for session ${session.id}`
-        );
-        session.proxyManager.removeListener('dry-run-complete', handler);
-      }
-      if (timeoutId) {
-        clearTimeout(timeoutId);
-      }
-    }
-  }
-
   async startDebugging(
     sessionId: string,
     scriptPath: string,
@@ -227,571 +164,22 @@ export abstract class SessionManagerOperations extends SessionManagerData {
     adapterLaunchConfig?: Record<string, unknown>,
     breakOnExceptions?: ExceptionBreakMode
   ): Promise<DebugResult> {
-    const session = this._getSessionById(sessionId);
-    this.logger.info(
-      `Attempting to start debugging for session ${sessionId}, script: ${scriptPath}, dryRunSpawn: ${dryRunSpawn}, dapLaunchArgs:`,
-      sanitizePayloadForLogging(dapLaunchArgs)
+    return this.launcher.startDebugging(
+      sessionId,
+      scriptPath,
+      scriptArgs,
+      dapLaunchArgs,
+      dryRunSpawn,
+      adapterLaunchConfig,
+      breakOnExceptions
     );
-
-    // Fail fast when the adapter is known-unavailable (issue #360): consult
-    // the same toolchain probe list_supported_languages reports, BEFORE any
-    // state mutation or proxy teardown, so the caller gets the real reason
-    // instead of success-then-silence. Fails open when the probe can't tell.
-    const launchGate = await checkLaunchToolchain(
-      session.language,
-      this.adapterRegistry,
-      this.launchValidationCache,
-      this.logger
-    );
-    if (!launchGate.available) {
-      const error = ErrorMessages.launchUnavailable(session.language, launchGate.reason);
-      this.logger.warn(`[SessionManager] ${error}`);
-      return { success: false, state: session.state, error };
-    }
-
-    if (session.proxyManager) {
-      // Session-preserving teardown: closeSession here used to REMOVE the
-      // session from the store, so the state update below threw
-      // SessionNotFoundError and the session was silently destroyed (#238).
-      this.logger.warn(
-        `[SessionManager] Session ${sessionId} already has an active proxy. Terminating before starting new.`
-      );
-      await this.stopProxyPreservingSession(session);
-    }
-
-    // Update to INITIALIZING state and set lifecycle to ACTIVE
-    this._updateSessionState(session, SessionState.INITIALIZING);
-
-    // Explicitly set lifecycle state to ACTIVE when starting debugging.
-    // attachMode is cleared: a launch supersedes any prior attach, and a
-    // sticky flag would wrongly refuse restart_debugging forever (#238).
-    this.sessionStore.update(sessionId, {
-      sessionLifecycle: SessionLifecycleState.ACTIVE,
-      attachMode: false,
-    });
-    this.logger.info(`[SessionManager] Session ${sessionId} lifecycle state set to ACTIVE`);
-
-    // Record the launch spec for restart_debugging BEFORE attempting the
-    // launch — a start that dies mid-way is still meaningfully replayable.
-    // Dry runs are not recorded: lastLaunch means "most recent real launch".
-    if (!dryRunSpawn) {
-      session.lastLaunch = {
-        scriptPath,
-        scriptArgs,
-        dapLaunchArgs,
-        adapterLaunchConfig,
-        breakOnExceptions,
-        launchedAt: Date.now(),
-      };
-    }
-
-    try {
-      // For dry run, start the proxy and wait for completion
-      if (dryRunSpawn) {
-        // Mark that we're setting up a dry run handler
-        const sessionWithSetup = session as ManagedSession & { _dryRunHandlerSetup?: boolean };
-        sessionWithSetup._dryRunHandlerSetup = true;
-
-        // Start the proxy manager
-        await this.proxyLauncher.start(session, scriptPath, scriptArgs, dapLaunchArgs, dryRunSpawn, adapterLaunchConfig);
-        this.logger.info(`[SessionManager] ProxyManager started for session ${sessionId}`);
-        
-        // Check if already completed before waiting
-        const refreshedSession = this._getSessionById(sessionId);
-        this.logger.info(`[SessionManager] Checking state after start: ${refreshedSession.state}`);
-        
-        const initialDryRunSnapshot = refreshedSession.proxyManager?.getDryRunSnapshot?.();
-        const dryRunAlreadyComplete =
-          refreshedSession.state === SessionState.STOPPED ||
-          refreshedSession.proxyManager?.hasDryRunCompleted?.() === true;
-
-        if (dryRunAlreadyComplete) {
-          this.logger.info(
-            `[SessionManager] Dry run already completed for session ${sessionId}`
-          );
-          delete sessionWithSetup._dryRunHandlerSetup;
-
-          return {
-            success: true,
-            state: SessionState.STOPPED,
-            data: {
-              dryRun: true,
-              message: 'Dry run spawn command logged by proxy.',
-              command: initialDryRunSnapshot?.command,
-              script: initialDryRunSnapshot?.script,
-            },
-          };
-        }
-
-        // Wait for completion with timeout
-        this.logger.info(
-          `[SessionManager] Waiting for dry run completion with timeout ${this.dryRunTimeoutMs}ms`
-        );
-        
-        const dryRunCompleted = await this.waitForDryRunCompletion(
-          refreshedSession,
-          this.dryRunTimeoutMs
-        );
-        delete sessionWithSetup._dryRunHandlerSetup;
-
-        const latestSessionState = this._getSessionById(sessionId);
-        const latestSnapshot =
-          latestSessionState.proxyManager?.getDryRunSnapshot?.() ?? initialDryRunSnapshot;
-        const effectiveDryRunComplete =
-          dryRunCompleted ||
-          latestSessionState.state === SessionState.STOPPED ||
-          latestSessionState.proxyManager?.hasDryRunCompleted?.() === true;
-
-        if (effectiveDryRunComplete) {
-          this.logger.info(
-            `[SessionManager] Dry run completed for session ${sessionId}, final state: ${latestSessionState.state}`
-          );
-
-          return {
-            success: true,
-            state: SessionState.STOPPED,
-            data: {
-              dryRun: true,
-              message: 'Dry run spawn command logged by proxy.',
-              command: latestSnapshot?.command,
-              script: latestSnapshot?.script,
-            },
-          };
-        } else {
-          // Timeout occurred
-          const finalSession = latestSessionState;
-          this.logger.error(
-            `[SessionManager] Dry run timeout for session ${sessionId}. ` +
-              `State: ${finalSession.state}, ProxyManager active: ${!!finalSession.proxyManager}`
-          );
-
-          return {
-            success: false,
-            error: `Dry run timed out after ${this.dryRunTimeoutMs}ms. Current state: ${finalSession.state}`,
-            state: finalSession.state,
-          };
-        }
-      }
-
-      // Normal (non-dry-run) flow
-      // Resolve the effective breakOnExceptions mode (issue #244): when the
-      // user did not specify one, launch sessions take the adapter policy's
-      // default. Attach-shaped configs are excluded — pausing a process you
-      // attached to on exceptions is surprising — as are dry runs (above).
-      const launchArgsShape = dapLaunchArgs as Record<string, unknown> | undefined;
-      const isAttachShaped =
-        launchArgsShape?.request === 'attach' || launchArgsShape?.__attachMode === true;
-      let effectiveBreakOnExceptions = breakOnExceptions;
-      if (effectiveBreakOnExceptions === undefined && !isAttachShaped) {
-        const policyDefault = this.selectPolicy(session.language)
-          .getInitializationBehavior?.().defaultExceptionBreakMode;
-        if (policyDefault) {
-          effectiveBreakOnExceptions = policyDefault;
-          this.logger.info(
-            `[SessionManager] Applying policy default breakOnExceptions='${policyDefault}' for ${session.language} launch session ${sessionId}`
-          );
-        }
-      }
-      session.effectiveBreakOnExceptions = effectiveBreakOnExceptions;
-
-      // Start the proxy manager
-      const launchConfigData = await this.proxyLauncher.start(session, scriptPath, scriptArgs, dapLaunchArgs, dryRunSpawn, adapterLaunchConfig, effectiveBreakOnExceptions);
-      this.logger.info(`[SessionManager] ProxyManager started for session ${sessionId}`);
-
-      // Perform language-specific handshake if required
-      const policy = this.selectPolicy(session.language);
-      if (policy.performHandshake) {
-        try {
-          await policy.performHandshake({
-            proxyManager: session.proxyManager,
-            sessionId: session.id,
-            dapLaunchArgs,
-            scriptPath,
-            scriptArgs,
-            breakpoints: session.breakpoints,
-            launchConfig: launchConfigData,
-            breakOnExceptions: effectiveBreakOnExceptions
-          });
-        } catch (handshakeErr) {
-          this.logger.warn(
-            `[SessionManager] Language handshake returned with warning/error: ${
-              handshakeErr instanceof Error ? handshakeErr.message : String(handshakeErr)
-            }`
-          );
-        }
-      }
-
-      // Use policy-defined readiness criteria when available.
-      const sessionStateAfterHandshake = this._getSessionById(sessionId).state;
-      const alreadyReady = policy.isSessionReady
-        ? policy.isSessionReady(sessionStateAfterHandshake, { stopOnEntry: dapLaunchArgs?.stopOnEntry })
-        : sessionStateAfterHandshake === SessionState.PAUSED;
-
-      if (!alreadyReady) {
-        // Wait for adapter to be configured, first stop event, or termination
-        const waitForReady = new Promise<void>((resolve) => {
-          let resolved = false;
-          // eslint-disable-next-line prefer-const -- assigned after cleanup/handlers are defined
-          let timeoutId: ReturnType<typeof setTimeout> | undefined;
-
-          const cleanup = () => {
-            if (timeoutId) clearTimeout(timeoutId);
-            session.proxyManager?.removeListener('stopped', handleStopped);
-            session.proxyManager?.removeListener('adapter-configured', handleConfigured);
-            session.proxyManager?.removeListener('terminated', handleTerminated);
-            session.proxyManager?.removeListener('exited', handleExited);
-            session.proxyManager?.removeListener('exit', handleExit);
-          };
-
-          const handleStopped = () => {
-            if (!resolved) {
-              resolved = true;
-              cleanup();
-              this.logger.info(`[SessionManager] Session ${sessionId} stopped on entry`);
-              resolve();
-            }
-          };
-
-          const handleConfigured = () => {
-            const readyOnRunning = policy.isSessionReady
-              ? policy.isSessionReady(SessionState.RUNNING, { stopOnEntry: dapLaunchArgs?.stopOnEntry })
-              : !dapLaunchArgs?.stopOnEntry;
-            if (!resolved && readyOnRunning) {
-              resolved = true;
-              cleanup();
-              this.logger.info(
-                `[SessionManager] Session ${sessionId} running (stopOnEntry=${dapLaunchArgs?.stopOnEntry ?? false})`
-              );
-              resolve();
-            }
-          };
-
-          const handleTerminated = () => {
-            if (!resolved) {
-              resolved = true;
-              cleanup();
-              this.logger.info(`[SessionManager] Session ${sessionId} terminated during startup`);
-              resolve();
-            }
-          };
-
-          const handleExited = () => {
-            if (!resolved) {
-              resolved = true;
-              cleanup();
-              this.logger.info(`[SessionManager] Session ${sessionId} exited during startup`);
-              resolve();
-            }
-          };
-
-          const handleExit = () => {
-            if (!resolved) {
-              resolved = true;
-              cleanup();
-              this.logger.info(`[SessionManager] Session ${sessionId} proxy exited during startup`);
-              resolve();
-            }
-          };
-
-          session.proxyManager?.once('stopped', handleStopped);
-          session.proxyManager?.once('adapter-configured', handleConfigured);
-          session.proxyManager?.once('terminated', handleTerminated);
-          session.proxyManager?.once('exited', handleExited);
-          session.proxyManager?.once('exit', handleExit);
-
-          // In case the adapter already reached the desired state before listeners were attached,
-          // perform a synchronous state check to avoid waiting for an event that already fired.
-          const currentState = this._getSessionById(sessionId).state;
-          const readyNow = policy.isSessionReady
-            ? policy.isSessionReady(currentState, { stopOnEntry: dapLaunchArgs?.stopOnEntry })
-            : currentState === SessionState.PAUSED;
-          if (readyNow) {
-            resolved = true;
-            cleanup();
-            resolve();
-            return;
-          }
-
-          // Also check if already terminated/stopped
-          if (currentState === SessionState.STOPPED || currentState === SessionState.ERROR) {
-            resolved = true;
-            cleanup();
-            this.logger.info(`[SessionManager] Session ${sessionId} already ${currentState} - skipping readiness wait`);
-            resolve();
-            return;
-          }
-
-          // Timeout after 30 seconds
-          timeoutId = setTimeout(() => {
-            if (!resolved) {
-              resolved = true;
-              cleanup();
-              this.logger.warn(ErrorMessages.adapterReadyTimeout(30));
-              resolve();
-            }
-          }, 30000);
-        });
-
-        await waitForReady;
-      } else {
-        this.logger.info(
-          `[SessionManager] Session ${sessionId} already ${sessionStateAfterHandshake} after handshake - skipping adapter readiness wait`
-        );
-      }
-
-      // Re-fetch session to get the most up-to-date state
-      const finalSession = this._getSessionById(sessionId);
-      const finalState = finalSession.state;
-
-      // Belt-and-braces re-sync (issues #236/#439): the worker forwards its
-      // initial setBreakpoints results via the breakpoints_synced status, so
-      // the store is normally already stamped — including for launches that
-      // are STOPPED by now (logpoint-only short programs), which this gated
-      // path can never help. A live re-sync still heals anything that
-      // changed between the snapshot and now, or a status lost to an IPC
-      // hiccup. Replace-all with the identical set is idempotent;
-      // syncBreakpointsForFile no-ops unless live.
-      if (finalSession.breakpoints.size > 0 &&
-          (finalState === SessionState.RUNNING || finalState === SessionState.PAUSED)) {
-        const files = [...new Set(Array.from(finalSession.breakpoints.values()).map(bp => bp.file))];
-        for (const file of files) {
-          await this.breakpoints.syncBreakpointsForFile(finalSession, file);
-        }
-      }
-      // Same re-sync for function breakpoints (issue #271 phase 3): the
-      // worker's initial send's responses never reach this store either.
-      if ((finalSession.functionBreakpoints?.size ?? 0) > 0 &&
-          (finalState === SessionState.RUNNING || finalState === SessionState.PAUSED)) {
-        await this.breakpoints.syncFunctionBreakpoints(finalSession);
-      }
-
-      // Unbound-at-launch warning (issue #308): the verified state is fresh
-      // after the re-sync above, so a name the adapter could not resolve is
-      // reported here instead of failing silently at "the program never
-      // stopped". Suppressed for bind-late adapters (js/java), where
-      // unverified-at-launch is the designed deferral path.
-      const fnBpWarning = this.breakpoints.functionBreakpointLaunchWarning(finalSession);
-
-      // Ran-to-completion with breakpoints that never bound (issue #467):
-      // state "stopped" where the caller expected "paused" is only
-      // explainable via list_breakpoints today — surface the stored
-      // per-breakpoint diagnostics right here where the caller is looking.
-      const unboundAtExitWarning =
-        finalState === SessionState.STOPPED
-          ? buildUnboundBreakpointExitWarning(finalSession)
-          : undefined;
-
-      // Logpoint-downgrade verdict (issue #469): the deferred set_breakpoint
-      // warning promised a launch-time answer — deliver it on this response.
-      const logpointWarning = buildLogpointDowngradeLaunchWarning(finalSession);
-
-      // Adapter degradation notes (issue #441) accumulate on the session as
-      // annotated output events arrive; joining here is best-effort — a note
-      // arriving after this return still lands in the output buffer as an
-      // attributed [mcp-debugger] Warning entry.
-      const launchWarning =
-        [fnBpWarning, logpointWarning, unboundAtExitWarning, ...(finalSession.adapterNotices ?? [])]
-          .filter(Boolean)
-          .join('; ') || undefined;
-
-      this.logger.info(
-        `[SessionManager] Debugging started for session ${sessionId}. State: ${finalState}`
-      );
-
-      return {
-        success: true,
-        state: finalState,
-        data: {
-          ...(launchWarning ? { warning: launchWarning } : {}),
-          message: `Debugging started for ${scriptPath}. Current state: ${finalState}`,
-          // Prefer the actual DAP stop reason (issue #214) — the first stop is
-          // not always a breakpoint (e.g. an uncaught exception before any
-          // breakpoint is hit). handleStopped records lastStop synchronously
-          // before every user-visible PAUSED transition, so PAUSED without
-          // lastStop is only the auto-continue transient (an entry stop being
-          // auto-continued); report 'unknown' rather than fabricating
-          // 'breakpoint' there (issue #255 residual).
-          reason:
-            finalState === SessionState.PAUSED
-              ? finalSession.lastStop?.reason ??
-                (dapLaunchArgs?.stopOnEntry ? 'entry' : 'unknown')
-              : undefined,
-          stopOnEntrySuccessful: !!dapLaunchArgs?.stopOnEntry && finalState === SessionState.PAUSED,
-        },
-      };
-    } catch (error) {
-      const diagnosticData = await logProxyFailure(
-        { logger: this.logger, fileSystem: this.fileSystem },
-        session,
-        error,
-        'startDebugging'
-      );
-
-      const errorMessage = error instanceof Error ? error.message : String(error);
-
-      const toolchainValidation =
-        (error as { toolchainValidation?: ToolchainValidationState })?.toolchainValidation ??
-        session.toolchainValidation;
-      const incompatibleToolchain =
-        Boolean(toolchainValidation) && toolchainValidation?.compatible === false;
-
-      if (incompatibleToolchain) {
-        this._updateSessionState(session, SessionState.CREATED);
-        this.sessionStore.update(sessionId, {
-          sessionLifecycle: SessionLifecycleState.CREATED,
-        });
-      } else {
-        this._updateSessionState(session, SessionState.ERROR);
-      }
-
-      if (session.proxyManager) {
-        await session.proxyManager.stop();
-        session.proxyManager = undefined;
-      }
-
-      // Normalize error identity for callers/tests
-      let errorType: string | undefined;
-      let errorCode: number | undefined;
-      if (error instanceof McpError) {
-        errorType = (error as McpError).constructor.name || 'McpError';
-        errorCode = (error as McpError).code as number | undefined;
-      } else if (error instanceof Error) {
-        errorType = error.constructor.name || 'Error';
-      }
-
-      if (incompatibleToolchain && toolchainValidation) {
-        const behavior = (toolchainValidation.behavior ?? 'warn').toLowerCase();
-        const canContinue = behavior !== 'error';
-        const updatedSession = this._getSessionById(sessionId);
-        return {
-          success: false,
-          error: 'MSVC_TOOLCHAIN_DETECTED',
-          state: updatedSession.state,
-          data: {
-            message: toolchainValidation.message ?? errorMessage,
-            toolchainValidation,
-          },
-          canContinue,
-          errorType,
-          errorCode,
-        };
-      }
-
-      return {
-        success: false,
-        error: errorMessage,
-        state: session.state,
-        errorType,
-        errorCode,
-        ...(Object.keys(diagnosticData).length > 0 ? { data: diagnosticData } : {})
-      };
-    }
   }
 
-
-  /** Sessions with a restart currently in flight (reentrancy guard, #238) */
-  private restartingSessions = new Set<string>();
-
   /**
-   * Restart the debuggee: terminate the current program (if any) and replay
-   * the last real launch with the same configuration. Breakpoints re-apply
-   * automatically via the initialBreakpoints snapshot; the output buffer
-   * starts fresh (read from since=0). Terminate+relaunch is used uniformly —
-   * no adapter advertises native DAP restart, and the spec blesses the
-   * emulation — so every launch-mode language works with no per-adapter
-   * wiring (issue #238).
+   * Restart the debuggee by replaying the last real launch (issue #238).
    */
   async restartDebugging(sessionId: string): Promise<DebugResult> {
-    const session = this._getSessionById(sessionId);
-
-    if (session.attachMode) {
-      return {
-        success: false,
-        state: session.state,
-        error: 'Cannot restart an attach session: there is no launch configuration to replay. Detach and re-attach instead.'
-      };
-    }
-    if (!session.lastLaunch) {
-      return {
-        success: false,
-        state: session.state,
-        error: 'Nothing to restart: this session has not been launched (start_debugging has not run, or only a dry run was performed).'
-      };
-    }
-    if (this.restartingSessions.has(sessionId)) {
-      return {
-        success: false,
-        state: session.state,
-        error: 'A restart is already in progress for this session.'
-      };
-    }
-    if (session.state === SessionState.INITIALIZING) {
-      return {
-        success: false,
-        state: session.state,
-        error: 'Session is still initializing; wait for the current start to complete before restarting.'
-      };
-    }
-
-    this.restartingSessions.add(sessionId);
-    try {
-      // Content anchors re-resolve BEFORE the relaunch snapshots
-      // initialBreakpoints, so breakpoints survive the edit that was the
-      // point of the session (issue #271).
-      const anchorResolution = await reresolveAnchors(session, this.opsContext);
-
-      const spec = session.lastLaunch;
-      this.logger.info(
-        `[SessionManager] Restarting session ${sessionId}: replaying launch of ${spec.scriptPath}`
-      );
-      const result = await this.startDebugging(
-        sessionId,
-        spec.scriptPath,
-        spec.scriptArgs,
-        spec.dapLaunchArgs,
-        false, // never replay as a dry run
-        spec.adapterLaunchConfig,
-        spec.breakOnExceptions
-      );
-      if (result.success) {
-        const staleCount = anchorResolution?.stale.length ?? 0;
-        // Stamp stale-anchor notes AFTER the relaunch: the per-launch
-        // breakpoint state reset (#238) clears message on every new launch,
-        // and a real adapter message should still win over ours.
-        if (anchorResolution) {
-          const bps = this._getSessionById(sessionId).breakpoints;
-          for (const staleEntry of anchorResolution.stale) {
-            const bp = bps.get(staleEntry.breakpointId);
-            if (bp && !bp.message) {
-              bp.message = `Anchor "${staleEntry.statement}" not found at restart; breakpoint kept at last known line ${staleEntry.line}`;
-            }
-          }
-        }
-        // Join rather than clobber: startDebugging may already have set a
-        // warning (unbound function breakpoints, issue #308).
-        const priorWarning = (result.data as { warning?: string } | undefined)?.warning;
-        const staleWarning = staleCount > 0
-          ? `${staleCount} statement anchor(s) no longer match the current file; those breakpoints kept their previous lines — re-set them if the target moved.`
-          : undefined;
-        const ambiguousCount = anchorResolution?.moved.filter((m) => m.candidates !== undefined).length ?? 0;
-        const ambiguousWarning = ambiguousCount > 0
-          ? `${ambiguousCount} statement anchor(s) matched multiple lines and re-anchored to the nearest match — check anchorResolution.moved (candidates listed) and re-set any that landed wrong.`
-          : undefined;
-        const warnings = [priorWarning, staleWarning, ambiguousWarning].filter(Boolean);
-        result.data = {
-          ...((result.data as object) ?? {}),
-          breakpointsReapplied: this._getSessionById(sessionId).breakpoints.size,
-          // Each launch starts a fresh output buffer: tell the caller to
-          // reset its get_output cursor to since=0.
-          outputReset: true,
-          ...(anchorResolution ? { anchorResolution } : {}),
-          ...(warnings.length > 0 ? { warning: warnings.join('; ') } : {}),
-        };
-      }
-      return result;
-    } finally {
-      this.restartingSessions.delete(sessionId);
-    }
+    return this.launcher.restartDebugging(sessionId);
   }
 
   /**
